Sorting: Bubble Sort
Given an -element array, , of distinct elements, sort array  in
ascending order using the Bubble Sort algorithm above. Once sorted,
print the following three lines:

Array is sorted in numSwaps swaps., where it is the number of swaps that
took place.
First Element: firstElement, where it is the first element in the sorted
array.
Last Element: lastElement, where it  is the last element in the sorted
array.